% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aco_rj.R
\name{aco_rj}
\alias{aco_rj}
\title{Implementation of ant colony optimization by ross jacobucci}
\usage{
aco_rj(size, iters, fitness, criterion)
}
\arguments{
\item{size}{}

\item{iters}{}
}
\description{
Implementation of ant colony optimization by ross jacobucci
}
\examples{
aco_rj()
}
\keyword{aco_rj}

