% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoSEM.R
\name{autoSEM}
\alias{autoSEM}
\title{This function houses a number of different heuristic optimization algorithms for specification search.}
\usage{
autoSEM(method = "tabuSearch", data = NULL, nfac = NULL, varList = NULL,
  criterion = "BIC", minInd = 3, stdlv = TRUE, orth = TRUE,
  niter = 30, parallel = "no", replaceSamp = FALSE)
}
\arguments{
\item{method}{which optimization algorithm to use}
}
\description{
This function houses a number of different heuristic optimization algorithms for specification search.
}
\examples{
autoSEM()
}
\keyword{autoSEM}

