% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fit_indices.R
\name{d}
\alias{d}
\title{Calculates the fit functions}
\usage{
d(chisq, df, N)
}
\arguments{
\item{chisq}{}

\item{df}{degrees of freedom}

\item{N}{sample size}

\item{ncp}{non-centrality parameter}
}
\description{
Calculates the fit functions
}
\examples{
d()
rmsea()
}
\keyword{chisq}
\keyword{fit}
\keyword{ncp}
\keyword{rmsea}

