% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.autoSEM}
\alias{summary.autoSEM}
\title{Summary results from autoSEM.}
\usage{
\method{summary}{autoSEM}(object, ...)
}
\arguments{
\item{object}{An object from autoSEM.}

\item{...}{Other arguments.}
}
\description{
Summary results from autoSEM.
}

