% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabu_rj.R
\name{tabu_rj}
\alias{tabu_rj}
\title{Implementation of tabu search by ross jacobucci}
\usage{
tabu_rj(size, iters, fitness)
}
\arguments{
\item{size}{Tabu list size. This is the number of variables * factors.}

\item{iters}{Maximum number of iterations to allow.}

\item{fitness}{Fitness function to use.}
}
\description{
Implementation of tabu search by ross jacobucci
}
\examples{
\dontrun{
tabu_rj()
}
}
\keyword{tabu_rj}

