% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aco_rj.R
\name{aco_rj}
\alias{aco_rj}
\title{Implementation of ant colony optimization by ross jacobucci}
\usage{
aco_rj(size, iters, fitness, criterion, min.improve, seed)
}
\arguments{
\item{size}{Tabu list size. This is the number of variables * factors.}

\item{iters}{Maximum number of iterations to allow.}

\item{fitness}{Fitness function to use.}

\item{criterion}{Criterion to use.}

\item{min.improve}{Number of iterations to wait for improvement
before breaking.}

\item{seed}{Random number seed.}
}
\description{
Implementation of ant colony optimization by ross jacobucci
}
\examples{
\dontrun{
aco_rj()
}
}
\keyword{aco_rj}

