% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/autoSEM.R
\name{autoSEM}
\alias{autoSEM}
\title{This function houses a number of different heuristic optimization algorithms
    for specification search.}
\usage{
autoSEM(method = "tabuSearch", data = NULL, nfac = NULL, varList = NULL,
  criterion = "BIC", minInd = 3, lav.control = list(orth = TRUE, std.lv =
  TRUE), niter = 30, parallel = "no", CV = "boot", R = 100,
  min.improve = niter, seed = NULL)
}
\arguments{
\item{method}{which optimization algorithm to use. Currently, it is only
       recommended to use "GA" for the genetic algorithm from the GA
       package, "aco_rj", an implementation of the ant colony
       algorithm by Ross Jacobucci, and "tabu_rj", an implementation of
       the Tabu search procedure by Ross Jacobucci. The latter two
       algorithms are based on the book chapter by Marcoulides &
       Leite, 2013. The other methods: "pso", "NMOF", "DEoptim",
       "tabuSearch", and "rgenoud" are all based on real-value
       optimization, not binary strings. This substantially increases
       the computation time and these methods are not currently
       recommended for use.}

\item{data}{a required dataset to search with.}

\item{nfac}{the number of factors to test.}

\item{varList}{list containing the names of the
variables to use from the dataset.}

\item{criterion}{The fit index to use as a criterion for
choosing the best model. Current options are "NCP",
"RMSEA","AIC", "BIC", and "BIC2", which is the sample
size adjusted BIC.}

\item{minInd}{The minimum number of indicators per factor.}

\item{lav.control}{Additional arguments to pass to cfa(). An example is
is setting lav.control = list(orth=FALSE,std.lv=TRUE).}

\item{niter}{The maximum number of iterations to all.}

\item{parallel}{Whether to use the snowfall package for parallelization.
Note that this is only applicable for the GA package at this time.}

\item{CV}{Whether to use cross-validation for choosing the best model. The
default is to use fit indices without CV.}

\item{R}{If using bootstrap, how many samples to take? Default is 100}

\item{min.improve}{Number of iterations to wait for improvement
before breaking.}

\item{seed}{random seed number.}
}
\description{
This function houses a number of different heuristic optimization algorithms
    for specification search.
}
\examples{
\dontrun{
autoSEM()
}
}
\keyword{autoSEM}

