% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multFac.R
\name{multFac}
\alias{multFac}
\title{Tests multiple factors}
\usage{
multFac(facList, parallel = "no", ncore = 1, method = "tabuSearch",
  data = NULL, varList = NULL, criterion = "BIC", minInd = 3,
  stdlv = TRUE, orth = TRUE, niter = 30, CV = FALSE)
}
\arguments{
\item{method}{}
}
\description{
Tests multiple factors
}
\examples{
multFac()
}
\keyword{multFac}

