% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tabu_rj.R
\name{tabu_rj}
\alias{tabu_rj}
\title{Implementation of tabu search by ross jacobucci}
\usage{
tabu_rj(size, iters, fitness)
}
\arguments{
\item{size}{}

\item{iters}{}
}
\description{
Implementation of tabu search by ross jacobucci
}
\examples{
tabu_rj()
}
\keyword{tabu_rj}

